Fix missing __hash__ of UnsetValue and Defaults
This fixes an incompatibility with Python 3.11, where the dataclass implementation has changed to check for the immutability of default values by checking if they are hashable. This wasn't true for UnsetValue, so it couldn't be used as a default value anymore.

For the Default classes, __hash__ isn't really necessary, but these objects still should be considered immutable.