#116: Silence "incompatible overrides" warnings for specialized validators
According to OOP principles, overriding the return type of validate() in a subclass in an incompatible way (e.g. date is not a subclass of str) is not allowed. However, with our validators it absolutely makes sense to use base validators and convert their output to different types (e.g. string to date).